v1.1/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
flatbuffers==20181003210633
fonttools==4.60.1
fsspec==2025.9.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
onnx==1.19.0
onnxruntime==1.23.0
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.33.1
protobuf==6.32.1
psutil==7.1.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.2
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
